requirements.txt
streamlit>=1.25.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
PyMuPDF>=1.22.5
python-docx>=0.8.11
markdown>=3.4.3
Jinja2>=3.1.2
reportlab>=4.0.4
fpdf>=1.7.2
tqdm>=4.65.0
scikit-learn>=1.2.2
numpy>=1.24.3
weasyprint>=59.0
pytest>=7.3.1